Add keywords to improve discoverability
It's difficult to find `electron-builder` on [npmjs.com](https://www.npmjs.com/). For example if you search for [electron installer](https://www.npmjs.com/search?q=electron+installer), it's not even listed.
I added a bunch of keywords that users might use when trying to find a solution for building Electron installers. It should improve the discoverability of this project.