feat: add console flat --ignore
A pattern which specifies which files to ignore when copying files to
create the package(s). The out directory is ignored by default, along
with the electron-prebuilt and electron-packager Node modules, the .git
directory, and node_modules/.bin. Alternatively, this can be a predicate
function that, given the file path, returns true if the file should be
ignored or false if the file should be kept.