Clarify that simple auto-update works on Windows
I was sent on a bit of a boondoggle by the sentence "simplified auto-update is not supported for Squirrel.Windows". I misinterpreted this to mean it wasn't supported on Windows at all. Perhaps I'm the only one dumb enough to make this mistake, but I wanted to clarify the sentence for those poor clods as thick headed as I am.