feat: Allow passing through binaries and requirements for special signing cases
In order to properly sign helpers, such as keytar (used for keychain access), the specific binary needs to be signed.
electron-osx-sign allows this but there was no way to pass these through, so allow the option(s) in package.json.
Additionally, the mac (not app store) apps need requirements sent to ensure helper acts are recognized properly.