fix(publish): cache publishers based on the target name
Publishers are currently cached by the publisher provider (e.g.
"bintray") and by the platform name (e.g. "linux"). This means that if
the user has different publish configurations for targets that belong to
the same OS (e.g. deb and rpm), then one will override the other on the
cache.

Fixes: https://github.com/electron-userland/electron-builder/issues/1951
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>