feat: twice smaller block map file (use md5 instead of sha256 and remove base64 padding)
protobuf size — BlockMap size: 153104, compressed: 151256 So, it means that it doesn't make sense - better to use deflate instead of complicating (another runtime dependency (google-protobuf), proto files and so on)

size encoding in a form where next value is a relative to previous doesn't make sense (zero savings in tests), since in our case next size can be less than previous (so, int will be negative and `-` symbol will be added)

DONE: sha2556 secure hash is not required, md5 collision-resistance is good for our purpose, secure hash algo not required, in any case sha512 checksum is checked for the whole file. And size of matched block is checked in addition to.