feat: use blake2s 18 output length to compute checksums
produced block map is not big as if using sha256, but still secure and offer acceptable collision resistance (in any case we also check block size if checksum matches)

Why 18 bytes? Because md5 (16 bytes) collision resistance is also acceptable for us, so, first number, divisible by 3 (to be base64 friendly), is used.