snap: update mime db and icon cache on install
In order to improve the startup time for snapped
applications we can generate the mime DB and the
icon caches at snap build time.
In order to do this, we can use snap composition to
override the desktop-gtk{2,3} part, adding an install
script which does this.
The remote part can't do this by default since it
could break legacy applications providing such files
but this won't be the case for electron apps.