feat(electron-updater): extensible sources
With this change, users can register their own provider factory
functions. This allows users to write their own auto update providers,
or to extend the ones provided by this library. To do this, a user calls
`providerFactory.registerProviderType` passing a name for the provider
type as a string, and a function which accepts a `data` argument of type
`PublishConfiguration | AllPublishOptions` and an instance of
`AppUpdater`, and returns an instance of `Provider<any>`.