stop nsis from uninstalling our application if `--package-file` does not exist.
when installing or upgrading an application, a command line
`--package-file` can be specified. when specified, this indicates that
a client wishes to use this package to install an app (if it exists).

let's say we have an application `starcall`.

when we run `install-v0.0.1.exe` to either install or upgrade
`starcall`, `electrion-builder` first uninstalls the existing version of
starcall, if any exists. only *after* we've uninstalled the previous
version do we check if the file at `--package-file` exists (and if it
exists, we install it). we'll go into why a `package.7z` might be passed
along but not exist later.

anyway - when that package doesn't exist, we fail silently and declare
that the installation is complete, even though we've just deleted
`starcall` from people's computers!! D:

this change makes nsis use `--package-file` if it exists, and if it
doesn't, falls back to installing `starcall-v0.0.1` from the internet.
that package url is version-specific, and declared in
`APP_PACKAGE_URL/APP_{ARCH}_NAME` (see webpackage.nsh).

this is the smallest, minimal fix to stop `electron-builder` completely
uninstalling an app during upgrade. other fixes require more reworking
of the nsis architecture, and that's much more involved than this.