fix(dmg): correctly encode Unicode characters in DMG licenses (#4428)
Fixes a bug where licenses containing characters that are outside the standard ASCII 7-bit character
set would not contain the correct content when added to macOS DMGs. The RTF escape sequence used to
encode these characters was missing the "placeholder" character that should be present after the
code point data. This commit adds a ? (question mark) character as the placeholder.