Allow Electron 6+ Renderer helper bundle ID to be specified
This is still a WIP, but this worked with my app when setting these
values in the "mac" section:

"helperBundleId": "io.hypertools.Command-E.helperBase",
"helperRendererBundleId": "io.hypertools.Command-E.helper",

There's still a bit of nastiness in that with this configuration the
other helpers (GPU, Plugin) get unnecessarily weird bundle IDs such as
io.hypertools.Command-E.helperBase, but I can clean up how these are
generated so it's more consistent. Maybe with Electron 6 it makes more
sense to allow users to specific `helperBundleIdBase`, which in my
case would be set to "io.hypertools.Command-E.helper" and allow
helpers with no override to have a sane, consistent base instead of
piggybacking on helperBundleId.

Let me know what you think. I did confirm that macOS permission pop up
key on the Bundle Identifier, not on name or anything else.