fix(nsis): Error handling if uninstaller cannot be launched (#4674)
This adds error handling if the old-uninstaller.exe process cannot be
launched. Previously the error flag was not checked after calling
ExecWait (ifErrors was missing).

In case old-uninstaller.exe cannot be launched from inside $PLUGINSDIR
(a.k.a a subfolder of the Temp directory) - for example due to a group policy
restriction that prevents executables from being launched that reside
inside the Temp folder - the original uninstaller residing inside the
application install directory is executed in place.

If this still fails the update is aborted.