feat(appx): add support for custom extensions in appmanifest.xml
Add option for specifying a path relative to the app root that gives the user the ability to specify custom xml that should be merged into the <Extensions/> node in the appmanifest.xml. This gives the user the ability to add custom extensions that don't make sense as individual features, without breaking existing support for the options that fall under the <Extensions/> node.