feat: allow customization of bundleIdentifier for the new Electron 6+ renderer processes (#4692)
* Allow Electron 6+ Renderer helper bundle ID to be specified

This is still a WIP, but this worked with my app when setting these
values in the "mac" section:

"helperBundleId": "io.hypertools.Command-E.helperBase",
"helperRendererBundleId": "io.hypertools.Command-E.helper",

There's still a bit of nastiness in that with this configuration the
other helpers (GPU, Plugin) get unnecessarily weird bundle IDs such as
io.hypertools.Command-E.helperBase, but I can clean up how these are
generated so it's more consistent. Maybe with Electron 6 it makes more
sense to allow users to specific `helperBundleIdBase`, which in my
case would be set to "io.hypertools.Command-E.helper" and allow
helpers with no override to have a sane, consistent base instead of
piggybacking on helperBundleId.

Let me know what you think. I did confirm that macOS permission pop up
key on the Bundle Identifier, not on name or anything else.

* Update comments, formatting to be consistent with project

* Keep project formatting consistent

* Fix typo

* Formatting

* Fix typo with relevant Electron versions

* Remove change to test file

* Update snapshot size properties

I based these updates on the expected output on this failing test:
https://app.circleci.com/jobs/github/electron-userland/electron-builder/2854

I presume my change affects some of the snapshots. Apologies if this
is not a kosher way to address the problem, I'm not well-versed in how
testing of electron-builder works.

* Update snapshot with new type error (added properties)

Close #4691