fix(auto-updater): Log the raw version string (#4947)
When the app updater fails because the version string can't be parsed, we are putting `latestVersion` into the logs, which we know is `null`. That's not too helpful, but putting the raw string from `updateInfo.version` is helpful because you might be able to see what's gone wrong with parsing it.