fix: Escape version in NSIS Updater during replace (#5655)
* fix: Escape version in NSIS Updater during replace

The updater derives the path to the old blockmap file by using the url of the new blockmap file and changing the version. The code uses a RegEx with the replace function in order to replace all occurrences of the version number.
The problem is that the version number is a string like '0.1.9-2+273e2eb' and contains characters like '.' and '+' that get interpreted as regex characters. This fixes that.