fix(nsis): retry closing the app until success
Do not continue installation while the app is running. The running
executable and dlls are locked until the app is closed and the 7z
extraction will result in a partial update. If app cannot be quit
forcibly - ask user to close it manually and retry the app check in
the loop until success.

This situation might happen when the app runs in elevated permissions
mode while the installer is started normally.