fix(nsis): generate uninstaller without elevating
In order to sign the uninstaller, we generate it by building the installer
in a special mode and then run it immediately to get the uninstaller stub
which we can then sign.  If the installer is set to require administrative
rights to run, this can lead to the build breaking as it would then require
administrative rights to generate the (to-be-signed) uninstaller.  We can
work around this by setting an environment variable to request Windows to
ignore the embedded manifest and always run (the stub installer) as the
invoker instead.