fix(builder-util): fix how env is passed to child processes
If env is passed in extraOptions, we should ensure that the overrides get
applied correctly to the child process.  We already take extraOptions.env
into account earlier when we set it up.  However, we need to ensure that
any overrides to stdio still apply to ensure we don't break the intention
from 14a7778dd9ed9a4a99fa1bf25ce7de1d064113cb.