Fix `null` error in proxy server logging
Node's HTTP server `address()` method returns `null` until the server is
listening for connections. This commit fixes a bug where `MacUpdater`
attempts to log the server port before it is listening. The existing log
event is still recorded, without the server address. When the server is
listening, a new event is recorded that includes the server address.