fix(msi): manually escape XML special characters when building XML
project.wxs contains invalid XML chars when building a package
containing an ampersand.

All attributes are already escaped thanks to EJS engine but the part
about files is built as a string and injected as-is in the template.

The next step would be to use an EJS template for the files too.

https://github.com/electron-userland/electron-builder/issues/5387