electron/universal has a new minimatch option 'x64ArchFiles'
Some projects need to ship x64 binaries which will be run under
Rosetta2 on arm64 Macs.

This option allows a packager to specify that they're aware that
certain files in their distribution which are included in both x64
and arm64 builds are actually x64 binaries in both builds.

Without this option, electron/universal will attempt to lipo
the two binaries together, resulting in a build failure.