Update snapcraft.yaml to use gnome-3-38-2004
The core18 and plug gnome-3-28-1804 lose standard (free/community) support in less than a year in accordance with the Ubuntu LTS timeline as these both correspond with Ubuntu 18.04 LTS, which loses standard support in April 2023 and moves to paid-only support. To continue to have secure builds for Electron apps, new snap builds should move to core20 and use plug gnome-3-38-2004.

As an added bonus, this will also allow Ubuntu and official Ubuntu flavors to seed Electron snaps built with core20 and gnome-3-38-2004 because gnome-3-38-2004 provides the proper facility for this process.