feat(msi): build emulated arm64 MSI installers
Build x64 MSI installers when targeting the arm64 architecture. This
results in an x64 MSI installer that installs an arm64 version of the
application.

An x64 MSI installer is an improvement because building arm64 MSI
targets errors with:
```
error CNDL0264 : The parameter 'arch' is missing or has an invalid value arm64.
Possible values are x86, x64, or ia64.
```

This is a stopgap until the electron-builder-binaries wix version is upgraded
to a version that supports arm64 which is currently blocked by a number of
breaking changes between 4.0.0.5512.2 and 4.0.1. This upgrade is
discussed in #6077.