fix: order files in asar for optimized differential updates (#8128)
ASAR file begins with a header that list all files and an offset to each
file in the rest of the file. When a file placed early in ASAR changes
its length - it means that all subsequent file declarations in the
header will have their offsets updated. While harmless by itself, this
negatively affects the incremental download size as more of the
installer binary is different from what it used to be.

In this change we order files in asar such that:

- Dependencies/node_modules come first (they change least often)
- Main app files come last (they change more frequently)

Additionally, files in asar are now ordered alphabetically within each
fileset to guarantee stable output.

All of above results in 2x improvement of incremental download size.