chore(docs): migrating to `typedoc` and upgrading `mkdocs`+`material` packages (#8491)
Huge lift to get the documentation site in a much better state.

Upgrading `mkdoc` and related theme packages. Migrated to `typedoc` for
a variety of reasons, most notably that it can provide more info of
properties when rendered (such as `@deprecated` messages)

Migrates `docs` dir to `pages` as template documents. We copy this over
to `docs` before generating `.md` with typedoc into `docs` dir. This is
all automatically handled when we run `pnpm docs:all` (if local) or `sh
./netlify-docs.sh` (if netlify)

Dynamically inserts `.md` files via `{% includes <path> %}` tags to
streamline maintaining parity with released state of typedoc.

Adjusts CSS to allow full width of browser window to be used (better
readability of docs)