fix(mac): only fuse macOS universal builds on the combined universal package
Without this `@electron/universal` would error complaining that the x64 and the arm64 builds varied in an unexpected place.
Could have maybe overridden `addElectronFuses` in MacPackager, but we didn't have a reference to the fact we were building a universal package from the param in `doPack` and didn't want to explore adding more properties in the area.

Fixes https://github.com/electron-userland/electron-builder/issues/8798