feat: customNsisResources override in nsis options
Prior to this change `customNsisBinary` could be used to override the
url of `nsis-*.7z` from the `electron-builder-binaries` repo to a custom
location. However, there was no option to change the url of the
`nsis-resources-*.7z` file.

With this chcange the url of the NSIS resources could also be overriden.